+++ website/content/dashboard.php
@@ -34,6 +34,19 @@
                 <button class="input-button" id="dayTry">Try</button>
             </div>
         </div>
+        <div class="problem" id="resources">
+            <div class="content">
+                <div class="content-1">
+                    <h2>Resources</h2>
+                    <div id="resourcesLinks">
+                        <iframe width="560" height="315" src="https://www.youtube.com/embed/mJwfpcXwYRU?si=VqFZVFNc7XKlKtS7" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe>
+                        <iframe width="560" height="315" src="https://www.youtube.com/embed/K8CQbSD9wis?si=fV0ASpKremmOlBIS" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe>
+                        <iframe width="560" height="315" src="https://www.youtube.com/embed/K8CQbSD9wis?si=fV0ASpKremmOlBIS" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe>
+                        <iframe width="560" height="315" src="https://www.youtube.com/embed/K8CQbSD9wis?si=fV0ASpKremmOlBIS" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe>
+                    </div>
+                </div>
+            </div>
+        </div>
     </div>
     <!-- SCRIPT -->
     <script src="../script/dashboard.js"></script>