+++ website/modules/js/user.js
@@ -1,7 +1,13 @@
 // CONNECT TO ELEMENTS
 const userName = document.getElementById("userName");
 const userInsight = document.getElementById("userInsight");
+const preferencesLogo = document.getElementById("preferencesLogo");
+
+// PREFERENCES REDIRECTS TO PREFERENCES
+preferencesLogo.addEventListener("click", function() {
+    redirect("preferences")
+})
 
 // MODIFY VALUES
-userName.innerText = gc("userName")
-userInsight.innerText = gc("userInsight")
+userName.innerText = getCookie("userName")
+userInsight.innerText = getCookie("userInsight")