+++ website/modules/navbar.php
@@ -1,6 +1,6 @@
 <!-- NAVBAR -->
 <div id="navbar">
-    <img src="../public/logo.svg" alt="Abscissa logo" id="navbarLogo" class="navbarIcon">
-    <img src="../public/search.svg" alt="Search button" id="navbarSearch" class="navbarIcon">
+    <img src="../public/logo.svg" alt="Abscissa logo" id="navbarLogo" class="navbar-icon">
+    <img src="../public/search.svg" alt="Search button" id="navbarSearch" class="navbar-icon">
 </div>
 <script src="../modules/js/navbar.js"></script>