+++ website/modules/user.php
@@ -1,7 +1,7 @@
-<!-- USER UI INTEGRATION -->
+<!-- USER UI -->
 <div id="userData">
     <p id="userInsight"></p>
     <p id="userName"></p>
-    <img src="" alt=""> <!-- PREFERENCES ICON -->
+    <img src="../public/preferences.svg" alt="" class="user-icon" id="preferencesLogo">
 </div>
 <script src="../modules/js/user.js"></script>